Revert extracting PrefixedKey
It causes a bug in MVCC2PLTest: testRead
